% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{computeIMEC}
\alias{computeIMEC}
\title{Computes the Ising Model of Explanatory Coherence.}
\usage{
computeIMEC(
  matrix,
  evidence,
  phenomena,
  theory1,
  theory2 = character(),
  analytic = T
)
}
\arguments{
\item{matrix}{matrix of explanatory relations.}

\item{evidence}{vector of evidence for phenomena.}

\item{phenomena}{vector of phenomena should be the same length as evidence.}

\item{theory1}{vector of propositions in theory1.}

\item{theory2}{vector of propositions in theory2.}

\item{analytic}{whether the result should be calculated analytically or (for large networks) estimated using
Metropolis-Hastings algorithm enhanced with Coupling from the past (Murray, 2007).}
}
\value{
returns an IMEC object which contains the explanatory coherence of the propositions, the explanatory relations,
the evidence, and the phenomena
}
\description{
Computes IMEC based on previously specified explanatory relations.
}
\examples{
#Comparison of Oxygen and Phlogiston Theory of Combustion
Phenomena <- c("E1", "E2", "E3", "E4", "E5", "E6", "E7", "E8")
Oxygen <- c("OH1", "OH2", "OH3", "OH4", "OH5", "OH6")
Phlogiston <- c("PH1", "PH2", "PH3", "PH4", "PH5", "PH6")
evidence <- c(rep(1,8))
## oxygen and phlogiston
explanations <- initializeNetwork(Phenomena, Oxygen, Phlogiston)
#oygen explanations
explanations <- Explain(c("OH1", "OH2", "OH3"), "E1", explanations)
explanations <- Explain(c("OH1", "OH3"), "E3", explanations)
explanations <- Explain(c("OH1", "OH3", "OH4"), "E4", explanations)
explanations <- Explain(c("OH1", "OH5"), "E5", explanations)
explanations <- Explain(c("OH1", "OH4", "OH5"), "E6", explanations)
explanations <- Explain(c("OH1", "OH5"), "E7", explanations)
explanations <- Explain(c("OH1", "OH6"), "E8", explanations)
#phlogiston explanations
explanations <- Explain(c("PH1", "PH2", "PH3"), "E1", explanations)
explanations <- Explain(c("PH1", "PH3", "PH4"), "E2", explanations)
explanations <- Explain(c("PH5", "PH6"), "E5", explanations)
#Contradictions
explanations <- Contradict("PH3", "OH3", explanations)
explanations <- Contradict("PH6", "OH5", explanations)
IMEC <- computeIMEC(explanations,evidence, Phenomena, Oxygen, Phlogiston)
plot(IMEC)
summary(IMEC)

}
