% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMEC.R
\docType{package}
\name{IMEC}
\alias{IMEC}
\title{IMEC}
\description{
This package computes the Ising Model of Explanatory Coherence for theory comparison and
theory appraisal.
}
\section{Construct Explanary Network}{

\emph{intializeNetwork} constructs an initial empty explanatoy network
\emph{Explain} and \emph{Contradict} specify explanatory relations.
}

\section{Calculate IMEC}{

\emph{computeIMEC} computes the Ising model of explanatory coherence and returns
an object of class IMEC. Use summary to summarize the result and plot to plot the
explanatory relations.
}

