% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.IMEC}
\alias{plot.IMEC}
\title{Plots the explanatory relations}
\usage{
\method{plot}{IMEC}(x, nodesize = 10, ...)
}
\arguments{
\item{x}{Object of the class IMEC as returned by computeIMEC}

\item{nodesize}{size of vertices in the plotted network}

\item{...}{other parameters passed on to S3 method.}
}
\description{
Plot the explanatory relations between data and phenomena. A window will open where you
can drag the nodes in the intended position. Then press enter to plot the network.
}
\examples{
#Comparison of oxygen and phlogiston theory of combustion
phenomena <- c("E1", "E2", "E3", "E4", "E5", "E6", "E7", "E8")
oxygen <- c("OH1", "OH2", "OH3", "OH4", "OH5", "OH6")
phlogiston <- c("PH1", "PH2", "PH3", "PH4", "PH5", "PH6")
evidence <- c(rep(1,8))
## oxygen and phlogiston
explanations <- initializeNetwork(phenomena, oxygen, phlogiston)
#oygen explanations
explanations <- explain(c("OH1", "OH2", "OH3"), "E1", explanations)
explanations <- explain(c("OH1", "OH3"), "E3", explanations)
explanations <- explain(c("OH1", "OH3", "OH4"), "E4", explanations)
explanations <- explain(c("OH1", "OH5"), "E5", explanations)
explanations <- explain(c("OH1", "OH4", "OH5"), "E6", explanations)
explanations <- explain(c("OH1", "OH5"), "E7", explanations)
explanations <- explain(c("OH1", "OH6"), "E8", explanations)
#phlogiston explanations
explanations <- explain(c("PH1", "PH2", "PH3"), "E1", explanations)
explanations <- explain(c("PH1", "PH3", "PH4"), "E2", explanations)
explanations <- explain(c("PH5", "PH6"), "E5", explanations)
#contradictions
explanations <- contradict("PH3", "OH3", explanations)
explanations <- contradict("PH6", "OH5", explanations)
IMEC <- computeIMEC(explanations,evidence, phenomena, oxygen, phlogiston)
plot(IMEC)
summary(IMEC)

}
