% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{computeIMEC}
\alias{computeIMEC}
\title{Computes the ising model of explanatory coherence.}
\usage{
computeIMEC(
  matrix,
  evidence,
  phenomena,
  theory1,
  theory2 = character(),
  analytic = T,
  analogy = numeric()
)
}
\arguments{
\item{matrix}{matrix of explanatory relations.}

\item{evidence}{vector of evidence for phenomena.}

\item{phenomena}{vector of phenomena should be the same length as evidence.}

\item{theory1}{vector of propositions in theory1.}

\item{theory2}{vector of propositions in theory2.}

\item{analytic}{whether the result should be calculated analytically or (for large networks) estimated using
Metropolis-Hastings algorithm enhanced with Coupling from the past (Murray, 2007).}

\item{analogy}{this argument is only for purposes of adding analogy in the future and should currently not be used.}
}
\value{
returns an IMEC object which contains the explanatory coherence of the propositions, the explanatory relations,
the evidence, and the phenomena
}
\description{
Computes IMEC based on previously specified explanatory relations.
}
\examples{
#Comparison of oxygen and phlogiston theory of combustion
phenomena <- c("E1", "E2", "E3", "E4", "E5", "E6", "E7", "E8")
oxygen <- c("OH1", "OH2", "OH3", "OH4", "OH5", "OH6")
phlogiston <- c("PH1", "PH2", "PH3", "PH4", "PH5", "PH6")
evidence <- c(rep(1,8))
## oxygen and phlogiston
explanations <- initializeNetwork(phenomena, oxygen, phlogiston)
#oygen explanations
explanations <- explain(c("OH1", "OH2", "OH3"), "E1", explanations)
explanations <- explain(c("OH1", "OH3"), "E3", explanations)
explanations <- explain(c("OH1", "OH3", "OH4"), "E4", explanations)
explanations <- explain(c("OH1", "OH5"), "E5", explanations)
explanations <- explain(c("OH1", "OH4", "OH5"), "E6", explanations)
explanations <- explain(c("OH1", "OH5"), "E7", explanations)
explanations <- explain(c("OH1", "OH6"), "E8", explanations)
#phlogiston explanations
explanations <- explain(c("PH1", "PH2", "PH3"), "E1", explanations)
explanations <- explain(c("PH1", "PH3", "PH4"), "E2", explanations)
explanations <- explain(c("PH5", "PH6"), "E5", explanations)
#contradictions
explanations <- contradict("PH3", "OH3", explanations)
explanations <- contradict("PH6", "OH5", explanations)
IMEC <- computeIMEC(explanations,evidence, phenomena, oxygen, phlogiston)
plot(IMEC)
summary(IMEC)

}
