% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{summary.IMEC}
\alias{summary.IMEC}
\title{Summary of an IMEC object.}
\usage{
\method{summary}{IMEC}(object, ...)
}
\arguments{
\item{object}{IMEC object.}

\item{...}{other paremeters passed on from S3 method.}
}
\description{
Summary of an IMEC object.
}
\examples{
#Comparison of Oxygen and Phlogiston Theory of Combustion
Phenomena <- c("E1", "E2", "E3", "E4", "E5", "E6", "E7", "E8")
Oxygen <- c("OH1", "OH2", "OH3", "OH4", "OH5", "OH6")
Phlogiston <- c("PH1", "PH2", "PH3", "PH4", "PH5", "PH6")
evidence <- c(rep(1,8))
## oxygen and phlogiston
explanations <- initializeNetwork(Phenomena, Oxygen, Phlogiston)
#oygen explanations
explanations <- explain(c("OH1", "OH2", "OH3"), "E1", explanations)
explanations <- explain(c("OH1", "OH3"), "E3", explanations)
explanations <- explain(c("OH1", "OH3", "OH4"), "E4", explanations)
explanations <- explain(c("OH1", "OH5"), "E5", explanations)
explanations <- explain(c("OH1", "OH4", "OH5"), "E6", explanations)
explanations <- explain(c("OH1", "OH5"), "E7", explanations)
explanations <- explain(c("OH1", "OH6"), "E8", explanations)
#phlogiston explanations
explanations <- explain(c("PH1", "PH2", "PH3"), "E1", explanations)
explanations <- explain(c("PH1", "PH3", "PH4"), "E2", explanations)
explanations <- explain(c("PH5", "PH6"), "E5", explanations)
#contradictions
explanations <- contradict("PH3", "OH3", explanations)
explanations <- contradict("PH6", "OH5", explanations)
IMEC <- computeIMEC(explanations,evidence, Phenomena, Oxygen, Phlogiston)
plot(IMEC)
summary(IMEC)

}
